requirements.txt
Flask==3.0.3
mediapipe==0.10.14
moviepy==1.0.3
numpy==2.1.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
scikit_learn==1.5.2
tensorflow==2.17.0
tensorflow_intel==2.17.0
